I HATE that 0xFFFFFFFF is bigger than Integer.MAX_VALUE

+++ src/main/java/edu/ksu/operatingsystems/javaos/cpu/Cpu.java
@@ -12,7 +12,7 @@ public interface Cpu {
     /**
      * @return The cpu's registers
      */
-    int[] getRegisters();
+    long[] getRegisters();
 
     /**
      * Set the process that is loaded into the registers
